refactor(components): 合并组件并整理导入结构
- 将 components/app.tsx 内容合并到 components/chart.tsx

- App 组件重命名为 ChartApp，更准确反映功能职责

- 整理 chart.tsx 导入结构：按第三方库、ECharts、项目模块分类

- 更新 content.ts 中的组件导入路径

- 删除冗余的 app.tsx 文件，简化项目结构
